+++ app/notifications/validators.py
@@ -21,7 +21,7 @@ from app.notifications.process_notifications import create_content_for_notificat
 from app.utils import get_public_notify_type_text
 from app.dao.service_email_reply_to_dao import dao_get_reply_to_by_id
 from app.dao.service_letter_contact_dao import dao_get_letter_contact_by_id
-from app.json_models import TemplateJSONModel
+from app.serialised_models import TemplateJSONModel
 
 from gds_metrics.metrics import Histogram
